backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
django-filter==24.2
django-cors-headers~=4.4.0
gunicorn~=22.0.0
requests==2.32.3
django-cryptography
yarl